Refactor: Object Stream
